fix: 修复编译错误
- 移除 service/scheduler/job_center.go 中重复定义的 TaskFilter 结构体
- 修复 service/cost/k8s.go 中 float 到 int64 的类型转换问题

+++ server/service/cost/k8s.go
@@ -458,7 +458,7 @@ func (s *K8sCostService) RightSizeWorkload(ctx context.Context, clusterID, names
 
         // 历史数据
         recommendation.UsageHistory = []UsageData{
-                {Date: "2024-01-01", CPU: 2.1, Memory: int64(4.5 * 1024 * 1024 * 1024)},
+                {Date: "2024-01-01", CPU: 2.1, Memory: int64(4.5 * float64(1024*1024*1024))},
                 {Date: "2024-01-02", CPU: 2.3, Memory: int64(5 * 1024 * 1024 * 1024)},
                 {Date: "2024-01-03", CPU: 2.0, Memory: int64(4 * 1024 * 1024 * 1024)},
         }
